+++ src/actions/index.js
@@ -14,7 +14,6 @@
 import { getSearchFilters, GET_SEARCH_FILTERS } from './getSearchFilters';
 import { getSearchResults, GET_SEARCH_RESULTS } from './getSearchResults';
 import { getCalendarResults, GET_CALENDAR_RESULTS } from './calendarSearch';
-import { getEventSearchResults, GET_EVENTSEARCH_RESULTS } from './eventSearch';
 
 import { setOriginalQuery, ORIGINAL_QUERY } from './setOriginalQuery';
 
@@ -47,8 +46,6 @@ export {
   GET_DETTAGLI_PROCEDIMENTO,
   getTwitterPosts,
   GET_TWITTER_POSTS,
-  getEventSearchResults,
-  GET_EVENTSEARCH_RESULTS,
   setOriginalQuery,
   ORIGINAL_QUERY,
 };