+++ src/components/ItaliaTheme/Blocks/Calendar/Item.jsx
@@ -2,14 +2,29 @@
 import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
+import { defineMessages, useIntl } from 'react-intl';
 import { Spinner } from 'design-react-kit/dist/design-react-kit';
+import cx from 'classnames';
 import { getCalendarDayResults } from '@italia/actions';
 import { UniversalLink } from '@plone/volto/components';
 import { flattenToAppURL } from '@plone/volto/helpers';
-import { useIntl } from 'react-intl';
-import cx from 'classnames';
 
 import { viewDate } from '@italia/helpers';
+
+const messages = defineMessages({
+  scadenza_bando: {
+    id: 'scadenza_bando',
+    defaultMessage: 'Scadenza dei termini per partecipare al bando',
+  },
+  scadenza_domande_bando: {
+    id: 'scadenza_domande_bando',
+    defaultMessage: 'Termine per le richieste di chiarimenti',
+  },
+  chiusura_procedimento_bando: {
+    id: 'chiusura_procedimento_bando',
+    defaultMessage: 'Chiusura del procedimento',
+  },
+});
 
 const Item = ({ day, path, data, inEdit }) => {
   const intl = useIntl();
@@ -45,7 +60,7 @@ const Item = ({ day, path, data, inEdit }) => {
           querystringResults.subrequests &&
           querystringResults.subrequests[day]?.items[day]?.map(
             (item, index) => (
-              <div key={index}>
+              <div key={index} div className="calendar-item">
                 <div className="pl-3">
                   <div>{item?.type}</div>
                   <UniversalLink
@@ -54,6 +69,28 @@ const Item = ({ day, path, data, inEdit }) => {
                   >
                     {item.title}
                   </UniversalLink>
+
+                  {item.scadenza_domande_bando &&
+                    _day.diff(item.scadenza_domande_bando, 'day') === 0 && (
+                      <div className="scadenza_message">
+                        {intl.formatMessage(messages.scadenza_domande_bando)}
+                      </div>
+                    )}
+                  {item.scadenza_bando &&
+                    _day.diff(item.scadenza_bando, 'day') === 0 && (
+                      <div className="scadenza_message">
+                        {intl.formatMessage(messages.scadenza_bando)}
+                      </div>
+                    )}
+                  {item.chiusura_procedimento_bando &&
+                    _day.diff(item.chiusura_procedimento_bando, 'day') ===
+                      0 && (
+                      <div className="scadenza_message">
+                        {intl.formatMessage(
+                          messages.chiusura_procedimento_bando,
+                        )}
+                      </div>
+                    )}
                 </div>
                 <hr />
               </div>