+++ src/components/ItaliaTheme/View/Commons/SideMenu.jsx
@@ -1,3 +1,6 @@
+/* eslint-disable react-hooks/exhaustive-deps */
+/* eslint-disable jsx-a11y/anchor-is-valid */
+/* eslint-disable no-unused-expressions */
 import { defineMessages, useIntl } from 'react-intl';
 import React, { useEffect, useState } from 'react';
 import { throttle } from 'lodash';
@@ -126,7 +129,7 @@ const SideMenu = ({ data }) => {
 
   const handleClickAnchor = (id) => (e) => {
     e.preventDefault();
-    document.getElementById(id).scrollIntoView({
+    document.getElementById(id)?.scrollIntoView?.({
       behavior: 'smooth',
       block: 'start',
     });