+++ src/config/italiaConfig.js
@@ -65,6 +65,8 @@ import DettagliProcedimentiView from '@italia/components/ItaliaTheme/View/Traspa
 
 import CardWithImageTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/CardWithImageTemplate';
 import CardWithImageTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/CardWithImageTemplateSkeleton';
+import CardWithImageTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/CardWithImageTemplateOptions';
+
 import SmallBlockLinksTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/SmallBlockLinksTemplate';
 import SmallBlockLinksTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/SmallBlockLinksTemplateSkeleton';
 
@@ -83,8 +85,10 @@ import GridGalleryTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/G
 import GridGalleryTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/GridGalleryTemplateSkeleton';
 import RibbonCardTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/RibbonCardTemplate';
 import RibbonCardTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/RibbonCardTemplateSkeleton';
+import RibbonCardTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/RibbonCardTemplateOptions';
 import MapTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/MapTemplate';
 import MapTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/MapTemplateSkeleton';
+import MapTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions';
 
 import BandiInEvidenceTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/BandiInEvidenceTemplate';
 import BandiInEvidenceTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/BandiInEvidenceTemplateSkeleton';
@@ -184,7 +188,6 @@ export default function applyConfig(voltoConfig) {
       edit: SearchSectionsEdit,
       restricted: false,
       mostUsed: false,
-      blockHasOwnFocusManagement: true,
       security: {
         addPermission: [],
         view: [],
@@ -200,7 +203,6 @@ export default function applyConfig(voltoConfig) {
       edit: CalendarEdit,
       restricted: false,
       mostUsed: false,
-      blockHasOwnFocusManagement: true,
       security: {
         addPermission: [],
         view: [],
@@ -216,7 +218,6 @@ export default function applyConfig(voltoConfig) {
       edit: EventSearchEdit,
       restricted: false,
       mostUsed: false,
-      blockHasOwnFocusManagement: true,
       security: {
         addPermission: [],
         view: [],
@@ -232,7 +233,6 @@ export default function applyConfig(voltoConfig) {
       edit: ArgumentsInEvidenceEdit,
       restricted: false,
       mostUsed: false,
-      blockHasOwnFocusManagement: true,
       security: {
         addPermission: [],
         view: [],
@@ -375,6 +375,7 @@ export default function applyConfig(voltoConfig) {
           label: 'Card con immagine',
           template: CardWithImageTemplate,
           skeleton: CardWithImageTemplateSkeleton,
+          templateOptions: CardWithImageTemplateOptions,
         },
         inEvidenceTemplate: {
           label: 'In evidenza',
@@ -390,11 +391,13 @@ export default function applyConfig(voltoConfig) {
           label: 'Card con nastro',
           template: RibbonCardTemplate,
           skeleton: RibbonCardTemplateSkeleton,
+          templateOptions: RibbonCardTemplateOptions,
         },
         mapTemplate: {
           label: 'Mappa',
           template: MapTemplate,
           skeleton: MapTemplateSkeleton,
+          templateOptions: MapTemplateOptions,
         },
         smallBlockLinksTemplate: {
           label: 'Blocco link solo immagini',
@@ -431,6 +434,14 @@ export default function applyConfig(voltoConfig) {
           template: AmministrazioneTrasparenteTablesTemplate,
           skeleton: AmministrazioneTrasparenteTablesTemplateSkeleton,
         },
+
+        // ****** Example: ******
+        //template_name: {
+        //   label: 'Template label',
+        //   template: TemplateComponent,
+        //   skeleton: TemplateSkeletonComponent,
+        //   templateOptions: TemplateSidebarOptionsComponent
+        // },
       },
       listing_bg_colors: [], //{name:'blue', label:'Blu'},{name:'light-blue', label:'Light blue'},{name:'sidebar-background', label:'Grey'}
       listing_items_colors: [], //{name:'blue', label:'Blu'},{name:'light-blue', label:'Light blue'},{name:'sidebar-background', label:'Grey'}
@@ -466,6 +477,9 @@ export default function applyConfig(voltoConfig) {
       Subsite: faSitemapSVG,
       UnitaOrganizzativa: faBuildingSVG,
       Modulo: faFileDownloadSVG,
+    },
+    defaultExcludedFromSearch: {
+      portalTypes: ['Image', 'File'],
     },
     italiaThemeViewsConfig: {
       imagePosition: 'afterHeader', // possible values: afterHeader, documentBody
@@ -608,6 +622,7 @@ export default function applyConfig(voltoConfig) {
     requiredBlocks: {
       ...config.blocks.requiredBlocks.concat(...customRequiredBlocks),
     },
+    showEditBlocksInBabelView: true,
   };
   delete config.blocks.blocksConfig.listing.templates.imageGallery; // removes imageGallery volto template, because we have our photoGallery template