+++ src/customizations/volto/components/manage/Blocks/Listing/ListingStyle.jsx
@@ -3,15 +3,16 @@ import PropTypes from 'prop-types';
 import config from '@plone/volto/registry';
 import TemplateWidget from '@plone/volto/components/manage/Blocks/Listing/TemplateWidget';
 import SimpleCardTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/SimpleCardTemplateOptions';
-import RibbonCardTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/RibbonCardTemplateOptions';
-import MapTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions';
-import CardWithImageTemplateOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/CardWithImageTemplateOptions';
+
 import DefaultOptions from '@italia/components/ItaliaTheme/Blocks/Listing/Options/DefaultOptions';
 
 const ListingStyle = ({ data, block, onChangeBlock, required = false }) => {
   const templatesConfig = config.blocks.blocksConfig.listing.templates;
 
   if (templatesConfig && Object.keys(templatesConfig).length > 1) {
+    const TemplateOptions = data.template
+      ? templatesConfig[data.template].templateOptions
+      : null;
     return (
       <div className="sidebar-listing-data listing-style">
         <TemplateWidget
@@ -37,25 +38,13 @@ const ListingStyle = ({ data, block, onChangeBlock, required = false }) => {
             onChangeBlock={onChangeBlock}
           />
         )}
-        {data.template === 'ribbonCardTemplate' && (
-          <RibbonCardTemplateOptions
+
+        {TemplateOptions && (
+          <TemplateOptions
             data={data}
             block={block}
             onChangeBlock={onChangeBlock}
-          />
-        )}
-        {data.template === 'cardWithImageTemplate' && (
-          <CardWithImageTemplateOptions
-            data={data}
-            block={block}
-            onChangeBlock={onChangeBlock}
-          />
-        )}
-        {data.template === 'mapTemplate' && (
-          <MapTemplateOptions
-            data={data}
-            block={block}
-            onChangeBlock={onChangeBlock}
+            required={required}
           />
         )}
       </div>